ajout securite en back sur operationController (add, delete for connected user)

+++ BACK/app/controllers/operationController.js
@@ -1,4 +1,5 @@
 const Operation = require('../models/Operation');
+const Board = require('../models/Board');
 
 const operationController = {
 
@@ -18,12 +19,18 @@ const operationController = {
 
     addOperationForConnectedUser: async (req,res) => {
 
-        // sécurité BACK supplémentaire pour que l'on ne puisse changer seulement les planches nous appartenant
-        //if(req.session.user.id === board_id) {
-        const newOperation = new Operation(req.body);
+        const boardTest = Board.findOneByUser(req.session.user.id,req.body.board_id);
+        if(req.body.board_id === boardTest.board_id) {
+            const newOperation = new Operation(req.body);
+            await newOperation.save();
+            res.json("operation sauvegardé !");
+        } 
+        else{
+            res.json("Vous ne pouvez pas supprimé cette operation, ")
+        }
+        
         console.log("ma new operation ! ",newOperation);
-        await newOperation.save();
-        res.json("operation sauvegardé !");
+        
         //}
         //else {
         //    res.status(400).json("Vous ne pouvez modifier seulement les planches vous appartenant !");
@@ -32,11 +39,17 @@ const operationController = {
 
     deleteOperationForConnectedUser: async (req,res) => {
 
-    
-        //if() securite pour que l'on ne puisse supprimer seulement les planches nous appartenant
+
+        const allBoards = await Board.findAllForUser(req.session.user.id);
         const operation = new Operation(await Operation.findOne(req.params.id));
-        await operation.delete();
-        res.status(200).json("Operation supprimé !");
+        const board = await Board.findOneByUser(req.session.user.id,operation.board_id);
+        if(!board || (board === [] )) {
+            res.json("Vous ne pouvez pas supprimé cette planche ! elle ne vous appartient pas ou n'existe pas!");
+        }
+        else {
+            await operation.delete();
+            res.status(200).json("Operation supprimé !");
+        }
     },
 
     updateOperationForConnectedUser: async (req,res) => {
